fix: resolve all CI/CD errors across TypeScript, PHPStan, security, and test suites
- TypeScript: add CustomerNotification, OrderReturn, FlashSale, OrderReturnItem types;
  add tracking_url to Shipment; add returns to Order; fix Uint8Array cast;
  fix nullable type handling in notifications and flash-sale-countdown
- PHPStan: add @property annotations to 25+ Eloquent models; fix scope visibility
  (public not protected); fix missing PaymentService and WebPush classes; fix
  env() string cast in sentry.php; fix Auth facade usage; fix protected increment()
  calls; fix NewsletterCampaignNotification introLine→line; fix DashboardService
  signature and OrderStatusEnum constant; remove unused AsListener attribute;
  fix ApiCacheHeaders instanceof check; regenerate baseline
- Security: upgrade phpseclib from 3.0.50 to 3.0.51 (CVE-2026-40194)
- Tests: fix ModularBootTest to pass array to DashboardService::getStats()

+++ server/config/sentry.php
@@ -9,11 +9,11 @@ return [
 
     'environment' => env('APP_ENV', 'production'),
 
-    'sample_rate' => env('GLITCHTIP_SAMPLE_RATE', env('SENTRY_SAMPLE_RATE', 1.0)),
+    'sample_rate' => (float) env('GLITCHTIP_SAMPLE_RATE', env('SENTRY_SAMPLE_RATE', 1.0)),
 
-    'traces_sample_rate' => env('GLITCHTIP_TRACES_SAMPLE_RATE', env('SENTRY_TRACES_SAMPLE_RATE', 0.2)),
+    'traces_sample_rate' => (float) env('GLITCHTIP_TRACES_SAMPLE_RATE', env('SENTRY_TRACES_SAMPLE_RATE', 0.2)),
 
-    'profiles_sample_rate' => env('GLITCHTIP_PROFILES_SAMPLE_RATE', env('SENTRY_PROFILES_SAMPLE_RATE', 0.2)),
+    'profiles_sample_rate' => (float) env('GLITCHTIP_PROFILES_SAMPLE_RATE', env('SENTRY_PROFILES_SAMPLE_RATE', 0.2)),
 
     'breadcrumbs' => [
         'logs' => true,
